chore: separate describe and stringingfy into different files
